Fixed (new) compile warnings

+++ src/extras.c
@@ -673,7 +673,7 @@ int _papi_hwi_get_native_event_info(unsigned int EventCode, PAPI_event_info_t * 
 		if (retval == PAPI_OK || retval == PAPI_EBUF) {
 
 			/* Fill in the info structure */
-			info->event_code = EventCode;
+		  info->event_code = (int)EventCode;
 			retval = _papi_hwd[cidx]->ntv_code_to_descr(EventCode, info->long_descr, sizeof(info->long_descr));
 			if (retval == PAPI_OK || retval == PAPI_EBUF) {
 				info->short_descr[0] = '\0';
